# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk platform validasi hasil kerja AI oleh ahli manusia.",
        "title": "AIValid API",
        "termsOfService": "https://aivalid.id/rules-content",
        "contact": {
            "name": "AIValid Support",
            "url": "https://aivalid.id/contact-support"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "api.aivalid.id",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/sudo": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "End the current sudo session. If a sudo token is provided, it will be specifically revoked.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth-Sudo"
                ],
                "summary": "Revoke sudo mode",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sudo token",
                        "name": "X-Sudo-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerMessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sudo/extend": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Extend the current sudo session duration. Requires a valid sudo token in the X-Sudo-Token header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth-Sudo"
                ],
                "summary": "Extend sudo session",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sudo token",
                        "name": "X-Sudo-Token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SudoExtendResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sudo/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Check whether sudo mode is currently active for the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth-Sudo"
                ],
                "summary": "Check sudo mode status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Sudo token",
                        "name": "X-Sudo-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SudoStatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/sudo/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Re-authenticate to enter sudo mode. Requires password, and optionally TOTP code or backup code if 2FA is enabled.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth-Sudo"
                ],
                "summary": "Verify sudo mode",
                "parameters": [
                    {
                        "description": "Sudo credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SudoVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.SudoVerifyResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.SwaggerErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.SudoExtendResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.SudoStatusResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "expires_in": {
                    "description": "seconds remaining",
                    "type": "integer"
                },
                "is_active": {
                    "description": "Whether sudo mode is active",
                    "type": "boolean"
                },
                "requires_totp": {
                    "description": "Whether TOTP is needed for sudo",
                    "type": "boolean"
                }
            }
        },
        "dto.SudoVerifyRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "backup_code": {
                    "description": "Alternative to TOTP",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "totp_code": {
                    "description": "Required if TOTP enabled",
                    "type": "string"
                }
            }
        },
        "dto.SudoVerifyResponse": {
            "type": "object",
            "properties": {
                "expires_at": {
                    "type": "string"
                },
                "expires_in": {
                    "description": "seconds",
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "sudo_token": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SwaggerErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "AUTH001"
                },
                "details": {
                    "type": "string",
                    "example": "Detail tambahan"
                },
                "message": {
                    "type": "string",
                    "example": "Email atau password salah"
                }
            }
        },
...
        "handlers.SwaggerMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operasi berhasil"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin JWT token. Type \"Bearer\" followed by a space and admin JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}